Mark some types as internal and clean up.
