+++ base/src/main/java/io/spine/code/proto/FileDescriptors.java
@@ -21,15 +21,20 @@ package io.spine.code.proto;
 
 import com.google.protobuf.DescriptorProtos.FileDescriptorProto;
 import com.google.protobuf.DescriptorProtos.FileDescriptorSet;
+import io.spine.annotation.Internal;
 import io.spine.io.ResourceFiles;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.net.URL;
+import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
@@ -40,7 +45,10 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.collect.Sets.newHashSet;
 import static com.google.common.collect.Streams.stream;
+import static com.google.common.io.Files.createParentDirs;
 import static io.spine.option.Options.registry;
+import static io.spine.util.Exceptions.illegalArgumentWithCauseOf;
+import static io.spine.util.Exceptions.illegalStateWithCauseOf;
 import static io.spine.util.Exceptions.newIllegalStateException;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toSet;
@@ -52,7 +60,8 @@ import static java.util.stream.Collectors.toSet;
  * @author Alex Tymchenko
  * @author Dmytro Dashenkov
  */
-public class FileDescriptors {
+@Internal
+public final class FileDescriptors {
 
     /**
      * The name of the descriptor set file.
@@ -144,6 +153,30 @@ public class FileDescriptors {
     }
 
     /**
+     * Merges the contents of the given files into a single descriptor set.
+     *
+     * <p>This method assumes that all the given files exist and contain instances of
+     * {@link FileDescriptorSet} Protobuf message.
+     *
+     * @param files the file to merge
+     * @return the {@link MergedDescriptorSet}
+     */
+    public static MergedDescriptorSet merge(Collection<File> files) {
+        FileDescriptorSet merged = files
+                .stream()
+                .map(File::getPath)
+                .map(FileDescriptors::parse)
+                .flatMap(Collection::stream)
+                .distinct()
+                .reduce(FileDescriptorSet.newBuilder(),
+                        FileDescriptorSet.Builder::addFile,
+                        (right, left) -> right.addAllFile(left.getFileList()))
+                .build();
+        MergedDescriptorSet result = new MergedDescriptorSet(merged);
+        return result;
+    }
+
+    /**
      * Retrieves a {@link Predicate} on a given type {@code T}.
      *
      * <p>The predicate is satisfied (returns {@code true}) iff the result of applying the given
@@ -188,6 +221,46 @@ public class FileDescriptors {
      */
     public static Predicate<FileDescriptorProto> isNotGoogleProto() {
         return IsNotGoogleProto.PREDICATE;
+    }
+
+    /**
+     * A view on a {@link FileDescriptorSet} after merging.
+     */
+    public static final class MergedDescriptorSet {
+
+        private final FileDescriptorSet descriptorSet;
+
+        private MergedDescriptorSet(FileDescriptorSet descriptorSet) {
+            this.descriptorSet = descriptorSet;
+        }
+
+        /**
+         * Writes this descriptor set into the given file.
+         *
+         * <p>If the file exists, it will be overridden. Otherwise, the file (and all its parent
+         * directories if necessary) will be created.
+         *
+         * @param destination the file to write this descriptor set into
+         */
+        public void writeTo(File destination) {
+            checkNotNull(destination);
+            prepareFile(destination);
+            try (OutputStream out = new BufferedOutputStream(new FileOutputStream(destination))) {
+                descriptorSet.writeTo(out);
+            } catch (IOException e) {
+                throw illegalStateWithCauseOf(e);
+            }
+        }
+
+        private static void prepareFile(File destination) {
+            try {
+                destination.delete();
+                createParentDirs(destination);
+                destination.createNewFile();
+            } catch (IOException e) {
+                throw illegalArgumentWithCauseOf(e);
+            }
+        }
     }
 
     /**
