+++ tools/model-compiler/src/main/java/io/spine/tools/compiler/validation/ValidationRulesLookup.java
@@ -62,35 +62,34 @@ public final class ValidationRulesLookup {
     }
 
     public static void processDescriptorSetFile(File descriptorSetFile, File targetDir) {
-        final Logger log = log();
-
+        Logger log = log();
         log.debug("Validation rules lookup started.");
         findRulesAndWriteProperties(descriptorSetFile, targetDir);
         log.debug("Validation rules lookup complete.");
     }
 
     private static void findRulesAndWriteProperties(File setFile, File targetDir) {
-        final List<FileDescriptorProto> files = parseSkipStandard(setFile.getPath());
-        final List<MessageDeclaration> declarations = allThat(files, new IsValidationRule());
+        List<FileDescriptorProto> files = parseSkipStandard(setFile.getPath());
+        List<MessageDeclaration> declarations = allThat(files, new IsValidationRule());
         writeProperties(declarations, targetDir);
     }
 
     private static void writeProperties(Iterable<MessageDeclaration> ruleDeclarations,
                                         File targetDir) {
-        final Map<String, String> propsMap = newHashMap();
+        Map<String, String> propsMap = newHashMap();
         for (MessageDeclaration declaration : ruleDeclarations) {
-            final TypeName typeName = declaration.getTypeName();
-            final String ruleTargets =
+            TypeName typeName = declaration.getTypeName();
+            String ruleTargets =
                     Options.option(declaration.getMessage(), validationOf)
                            .orElseThrow(() -> newIllegalArgumentException(declaration.getTypeName()
                                                                                      .value()));
             propsMap.put(typeName.value(), ruleTargets);
         }
 
-        final String fileName = io.spine.validate.rules.ValidationRules.fileName();
+        String fileName = io.spine.validate.rules.ValidationRules.fileName();
         log().debug("Writing the validation rules description to {}/{}.",
                     targetDir, fileName);
-        final PropertiesWriter writer = new PropertiesWriter(targetDir.getAbsolutePath(), fileName);
+        PropertiesWriter writer = new PropertiesWriter(targetDir.getAbsolutePath(), fileName);
         writer.write(propsMap);
     }
